feature(SB-87): list of recipes shows locked status
